requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
alembic==1.18.3
    # via
    #   what-to-eat (pyproject.toml)
    #   mlflow
altair==6.0.0
    # via streamlit
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   starlette
    #   watchfiles
argon2-cffi==25.1.0
    # via what-to-eat (pyproject.toml)
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
bcrypt==5.0.0
    # via passlib
blinker==1.9.0
    # via
    #   flask
    #   streamlit
boto3==1.42.44
    # via what-to-eat (pyproject.toml)
botocore==1.42.44
    # via
    #   boto3
    #   s3transfer
cachetools==6.2.6
    # via
    #   mlflow-skinny
    #   mlflow-tracing
    #   streamlit
certifi==2026.1.4
    # via requests
cffi==2.0.0
    # via
    #   argon2-cffi-bindings
    #   cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   flask
    #   mlflow-skinny
    #   streamlit
    #   uvicorn
cloudpickle==3.1.2
    # via mlflow-skinny
contourpy==1.3.3
    # via matplotlib
cryptography==46.0.4
    # via
    #   google-auth
    #   mlflow
    #   python-jose
cycler==0.12.1
    # via matplotlib
databricks-sdk==0.85.0
    # via
    #   mlflow-skinny
    #   mlflow-tracing
dnspython==2.8.0
    # via email-validator
docker==7.1.0
    # via mlflow
ecdsa==0.19.1
    # via python-jose
email-validator==2.3.0
    # via pydantic
fastapi==0.128.3
    # via
    #   what-to-eat (pyproject.toml)
    #   mlflow-skinny
flask==3.1.2
    # via
    #   flask-cors
    #   mlflow
flask-cors==6.0.2
    # via mlflow
fonttools==4.61.1
    # via matplotlib
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via
    #   mlflow-skinny
    #   streamlit
google-auth==2.48.0
    # via databricks-sdk
graphene==3.4.3
    # via mlflow
graphql-core==3.2.7
    # via
    #   graphene
    #   graphql-relay
graphql-relay==3.2.0
    # via graphene
greenlet==3.3.1
    # via sqlalchemy
gunicorn==23.0.0
    # via mlflow
h11==0.16.0
    # via uvicorn
httptools==0.7.1
    # via uvicorn
huey==2.6.0
    # via mlflow
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   requests
importlib-metadata==8.7.1
    # via
    #   mlflow-skinny
    #   opentelemetry-api
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   altair
    #   flask
    #   pydeck
jmespath==1.1.0
    # via
    #   boto3
    #   botocore
joblib==1.5.3
    # via scikit-learn
jsonschema==4.26.0
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
kiwisolver==1.4.9
    # via matplotlib
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   mako
    #   werkzeug
matplotlib==3.10.8
    # via mlflow
mlflow==3.9.0
    # via what-to-eat (pyproject.toml)
mlflow-skinny==3.9.0
    # via mlflow
mlflow-tracing==3.9.0
    # via mlflow
narwhals==2.16.0
    # via altair
numpy==2.4.2
    # via
    #   contourpy
    #   matplotlib
    #   mlflow
    #   pandas
    #   pydeck
    #   scikit-learn
    #   scipy
    #   skops
    #   streamlit
opentelemetry-api==1.39.1
    # via
    #   mlflow-skinny
    #   mlflow-tracing
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-proto==1.39.1
    # via
    #   mlflow-skinny
    #   mlflow-tracing
opentelemetry-sdk==1.39.1
    # via
    #   mlflow-skinny
    #   mlflow-tracing
opentelemetry-semantic-conventions==0.60b1
    # via opentelemetry-sdk
packaging==25.0
    # via
    #   altair
    #   gunicorn
    #   matplotlib
    #   mlflow-skinny
    #   mlflow-tracing
    #   skops
    #   streamlit
pandas==2.3.3
    # via
    #   mlflow
    #   streamlit
passlib==1.7.4
    # via what-to-eat (pyproject.toml)
pillow==12.1.0
    # via
    #   matplotlib
    #   streamlit
prettytable==3.17.0
    # via skops
protobuf==6.33.5
    # via
    #   databricks-sdk
    #   mlflow-skinny
    #   mlflow-tracing
    #   opentelemetry-proto
    #   streamlit
psycopg2-binary==2.9.11
    # via what-to-eat (pyproject.toml)
pyarrow==22.0.0
    # via
    #   mlflow
    #   streamlit
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   python-jose
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   what-to-eat (pyproject.toml)
    #   fastapi
    #   mlflow-skinny
    #   mlflow-tracing
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via what-to-eat (pyproject.toml)
pydeck==0.9.1
    # via streamlit
pyparsing==3.3.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   graphene
    #   matplotlib
    #   pandas
python-dotenv==1.2.1
    # via
    #   mlflow-skinny
    #   pydantic-settings
    #   uvicorn
python-jose==3.5.0
    # via what-to-eat (pyproject.toml)
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   mlflow-skinny
    #   uvicorn
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   what-to-eat (pyproject.toml)
    #   databricks-sdk
    #   docker
    #   mlflow-skinny
    #   streamlit
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via
    #   google-auth
    #   python-jose
s3transfer==0.16.0
    # via boto3
scikit-learn==1.8.0
    # via
    #   mlflow
    #   skops
scipy==1.17.0
    # via
    #   mlflow
    #   scikit-learn
    #   skops
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
skops==0.13.0
    # via mlflow
smmap==5.0.2
    # via gitdb
sqlalchemy==2.0.46
    # via
    #   what-to-eat (pyproject.toml)
    #   alembic
    #   mlflow
sqlparse==0.5.5
    # via mlflow-skinny
starlette==0.52.1
    # via fastapi
streamlit==1.54.0
    # via what-to-eat (pyproject.toml)
tenacity==9.1.3
    # via streamlit
threadpoolctl==3.6.0
    # via scikit-learn
toml==0.10.2
    # via streamlit
tornado==6.5.4
    # via streamlit
typing-extensions==4.15.0
    # via
    #   alembic
    #   altair
    #   fastapi
    #   graphene
    #   mlflow-skinny
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-settings
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via
    #   botocore
    #   docker
    #   requests
uvicorn==0.40.0
    # via
    #   what-to-eat (pyproject.toml)
    #   mlflow-skinny
uvloop==0.22.1
    # via uvicorn
watchdog==6.0.0
    # via streamlit
watchfiles==1.1.1
    # via uvicorn
wcwidth==0.6.0
    # via prettytable
websockets==16.0
    # via uvicorn
werkzeug==3.1.5
    # via
    #   flask
    #   flask-cors
zipp==3.23.0
    # via importlib-metadata
